added remove property/object functionality in form submission to frontend
